Let user handle the session (save and load)
