+++ src/utils.ts
@@ -1,31 +1,7 @@
-import { readFileSync, writeFileSync } from "fs";
 import { WhatsAppClientConfig } from "./whatsapp/interfaces"
 
 const L = console.log
 const E = console.error
-
-function configLoad(file: string) {
-    const obj = JSON.parse(readFileSync(file, 'utf8'))
-    const cfg: WhatsAppClientConfig = Object.assign({}, obj)
-    cfg.serverSecret = Buffer.from(obj.serverSecret, 'base64')
-    cfg.aesKey = Buffer.from(obj.aesKey, 'base64')
-    cfg.macKey = Buffer.from(obj.macKey, 'base64')
-    Object.keys(obj.keys).forEach(
-        k => cfg.keys[k] = Buffer.from(obj.keys[k], 'base64')
-    )
-    return cfg
-}
-
-function configStore(file: string, cfg: WhatsAppClientConfig) {
-    const obj: any = Object.assign({}, cfg)
-    obj.serverSecret = cfg.serverSecret.toString('base64')
-    obj.aesKey = cfg.aesKey.toString('base64')
-    obj.macKey = cfg.macKey.toString('base64')
-    Object.keys(cfg.keys).forEach(
-        k => obj.keys[k] = cfg.keys[k].toString('base64')
-    )
-    return writeFileSync(file, JSON.stringify(obj, null, 2))
-}
 
 const Color = {
     c(s: string) { return `\x1b[36m${s}\x1b[0m` },
@@ -36,4 +12,4 @@ const Color = {
     b(s: string) { return `\x1b[34m${s}\x1b[0m` },
     w(s: string) { return `\x1b[37m${s}\x1b[0m` },
 }
-export { configStore, configLoad, Color, L, E }
+export { Color, L, E }
